cleaned code
