+++ controllers/contacts-controllers.js
@@ -1,0 +1,101 @@
+const contactsOperations = require("../models/contacts");
+
+const Joi = require("joi");
+
+const contactShema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().min(10).required(),
+});
+
+const getAllContacts = async (req, res, next) => {
+  try {
+    const contacts = await contactsOperations.listContacts();
+    return res.json(contacts);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const getById = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const result = await contactsOperations.getContactById(id);
+    if (!result) {
+      res.status(404).json({ message: "Not found" });
+      return;
+    }
+
+    res.json({
+      status: "success",
+      code: 200,
+      data: { result },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const addNewContact = async (req, res, next) => {
+  try {
+    const { error } = contactShema.validate(req.body);
+
+    if (error) {
+      error.status = 400;
+      throw error;
+    }
+    const result = await contactsOperations.addContact(req.body);
+    return res.status(201).json({
+      data: { result },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const deleteContactById = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const result = await contactsOperations.removeContact(id);
+
+    if (!result) {
+      res.status(404).json({ message: "Not found" });
+      return;
+    }
+
+    res.json({
+      message: "contact deleted",
+      data: { result },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const updateContactById = async (req, res, next) => {
+  try {
+    const { error } = contactShema.validate(req.body);
+
+    if (error) {
+      error.status = 400;
+      throw error;
+    }
+    const { id } = req.params;
+    const result = await contactsOperations.updateContact(id, req.body);
+    res.json({
+      status: "success",
+      code: 200,
+      data: { result },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = {
+  getAllContacts,
+  getById,
+  addNewContact,
+  deleteContactById,
+  updateContactById,
+};
