+++ models/contacts.json
@@ -54,12 +54,6 @@
     "phone": "(748) 206-2688"
   },
   {
-    "id": "a1a4e430-c4a4-4825-bce3-a430729b021a"
-  },
-  {
-    "id": "ba74028c-e575-4526-bacf-b3b37cf43622"
-  },
-  {
     "id": "3ab9808c-d55d-484f-b32e-3db36a82f5c3",
     "name": "Kate Ostin",
     "email": "kateO@gmail.com",
@@ -88,5 +82,11 @@
     "name": "Luke Lamar",
     "email": "lukeO@gmail.com",
     "phone": "(486) 914-5674"
+  },
+  {
+    "name": "Thomas Jacobs",
+    "email": "tommy@gmail.com",
+    "phone": "(486) 235-9874",
+    "id": "5e22f375-b454-424b-b3f5-594c2c0d462c"
   }
 ]
