+++ routes/api/contacts.js
@@ -1,99 +1,17 @@
 const express = require("express");
 
-const contactsOperations = require("../../models/contacts");
-
 const router = express.Router();
 
-const Joi = require("joi");
+const { contacts: ctrl } = require("../../controllers");
 
-const contactShema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().min(10).required(),
-});
+router.get("/", ctrl.getAllContacts);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactsOperations.listContacts();
-    return res.json(contacts);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:id", ctrl.getById);
 
-router.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsOperations.getContactById(id);
-    if (!result) {
-      res.status(404).json({ message: "Not found" });
-      return;
-    }
+router.post("/", ctrl.addNewContact);
 
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:id", ctrl.deleteContactById);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactShema.validate(req.body);
-
-    if (error) {
-      error.status = 400;
-      throw error;
-    }
-    const result = await contactsOperations.addContact(req.body);
-    return res.status(201).json({
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactsOperations.removeContact(id);
-
-    if (!result) {
-      res.status(404).json({ message: "Not found" });
-      return;
-    }
-
-    res.json({
-      message: "contact deleted",
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = contactShema.validate(req.body);
-
-    if (error) {
-      error.status = 400;
-      throw error;
-    }
-    const { id } = req.params;
-    const result = await contactsOperations.updateContact(id, req.body);
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:id", ctrl.updateContactById);
 
 module.exports = router;
